Create RFID Door Lock System
The Arduino based RFID door lock system is secure and responsive as compared to other systems. With the use of Arduino, it becomes much easy to design. It is simple and cost efficient project and can be used as a basic access control mechanism. Their main advantage is contactless communication.